requirements.txt
flask
mediapipe
numpy
pandas
opencv
tensorflow
pyaudio
SpeechRecognition
mysql-connector-python
opencv-python